qemu: Remove return value from 'qemuHotplugRemoveManagedPR'
The only place which actually checked the return value would skip code
e.g. to delete unused files or stop no longer used services. The rest of
the callers ignored the value.

As this is expected to be used on cleanup code paths which have no
possibility to report errors we should remove the return value
completely.

Signed-off-by: Peter Krempa <pkrempa@redhat.com>
Reviewed-by: Pavel Hrdina <phrdina@redhat.com>